skills/jenkins-skill/requirements.txt
# Jenkins Skill 依赖
python-jenkins>=1.8.0
requests>=2.28.0

# 开发依赖
pytest>=7.0.0
pytest-cov>=4.0.0
